update documentation and commands
