+++ Sources/PassKit/Models/Pass/PassBeacon.swift
@@ -1,5 +1,9 @@
 import Foundation
 
+// https://developer.apple.com/library/archive/documentation/UserExperience/Reference/PassKit_Bundle/Chapters/LowerLevel.html#//apple_ref/doc/uid/TP40012026-CH3-SW4
+/// Information about a location beacon
+///
+///  - Note: Available in iOS 7.0
 public struct PassBeacon: Codable {
     /// Major identifier of a Bluetooth Low Energy location beacon.
     public var major: UInt16?
@@ -10,6 +14,12 @@ public struct PassBeacon: Codable {
     /// Text displayed on the lock screen when the pass is currently relevant. For example, a description of the nearby location such as “Store nearby on 1st and Main.”
     public var relevantText: String?
     
+    /// Information about a location beacon
+    /// - Parameters:
+    ///   - major: Major identifier of a Bluetooth Low Energy location beacon
+    ///   - minor: Minor identifier of a Bluetooth Low Energy location beacon
+    ///   - proximityUUID: Unique identifier of a Bluetooth Low Energy location beacon
+    ///   - relevantText: Text displayed on the lock screen when the pass is currently relevant
     public init(
         major: UInt16? = nil,
         minor: UInt16? = nil,
