--- Sources/PassKit/Models/PassDataDetectorType.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-public enum PassDataDetectorType: String, Codable {
-    case phoneNumber = "PKDataDetectorTypePhoneNumber"
-    case link = "PKDataDetectorTypeLink"
-    case address = "PKDataDetectorTypeAddress"
-    case calendarEvent = "PKDataDetectorTypeCalendarEvent"
-}
