--- Sources/PassKit/Models/PassTextAlignment.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-public enum PassTextAlignment: String, Codable {
-    case left = "PKTextAlignmentLeft"
-    case center = "PKTextAlignmentCenter"
-    case right = "PKTextAlignmentRight"
-    case natural = "PKTextAlignmentNatural"
-}
